requirements.txt
requests
rich
python-dotenv
google-generativeai
discord-webhook
duckduckgo_search
openai
python-ffmpeg
